requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.7.0
pillow>=9.5.0
scikit-learn>=1.2.0
scipy>=1.10.0

# Augmentation
albumentations>=1.3.0

# Imbalanced Learning
imbalanced-learn>=0.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Metrics and Logging
scikit-image>=0.20.0
wandb>=0.15.0
tensorboard>=2.13.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
pathlib>=1.0.1

# Graph Neural Networks
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0

# Optional (for production)
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.0